// src/app/models/tarjeta.model.ts
import { Cliente } from './cliente.model';
import { Empresa } from './empresa.model';
export class Tarjeta {
	nroOrden: number;
	viaTransporte: string;
	fecha: Date;
	cliente: Cliente;
	peso: number;
	bultos: number;
	cif: number;
	nroDUA: string;
	declarante: Empresa;
	ubicacion: string;
	destino: string;
	nroDeclaracion: string;
	nroManifiesto: string;
	fechaNumeracion: Date;
	fechaVcto: Date;
	depTemporal: string;
	docTransporte: string;
	mercancia: string;
	observaciones: string;
	nroFactComercial: string;
	fechaFactComercial: Date;
	tiMercancia: string;
	clase: string;
	subpartidaNac: string;
	vapAvi: string;
	fob: number;
	flete: number;
	seguro: number;
	ajuste: number;
	total: number;
	nroCont40: number;
	nroCont20CargSuelta: number;
	levanteMercaderia: boolean;
	paraVoBo: boolean;
	factDolares: boolean;
}